New module Options for option parsing
